requirements.txt
numpy==1.26.4
tflite-runtime>=2.14,<2.17

# Vision + imaging
opencv-python==4.6.0.66
Pillow>=9.5,<11

# Utilities
PyYAML>=6.0.1
requests>=2.31
pyautogui>=0.9.53


pyodbc